refactor: separate national and manufacturer timeseries
